pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sekha-cli"
version = "0.1.0"
description = "Command-line interface for Sekha AI Memory Controller"
readme = "README.md"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Sekha Contributors"}
]
dependencies = [
    # Use either local path (dev) or git URL (CI/production)
    "sekha @ file:///${PROJECT_ROOT}/../sekha-python-sdk",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5"
]

[project.scripts]
sekha = "sekha_cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=sekha_cli --cov-report=term-missing --cov-report=html --cov-fail-under=80"

[tool.coverage.run]
source = ["sekha_cli"]
omit = ["*/tests/*", "*/test_*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]